+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Added possibility to extract translation messages from a file or files besides extracting from a directory
  * Added `TranslationsCacheWarmer` to create catalogues at warmup
+ * Added `TranslatorTrait` which provides shortcuts for Translator `trans()` and `transChoice()` methods
 
 2.6.0
 -----